split project into two modules
