+++ Task_2_3_1/.idea/gradle.xml
@@ -9,6 +9,8 @@
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/model" />
           </set>
         </option>
       </GradleProjectSettings>
