+++ Task_2_3_1/app/build.gradle
@@ -5,7 +5,7 @@ plugins {
     id 'jacoco'
 }
 
-group = 'ru.nsu.aeliseev2'
+group = 'ru.nsu.aeliseev2.task231'
 version = '1.0-SNAPSHOT'
 
 repositories {
@@ -13,6 +13,7 @@ repositories {
 }
 
 dependencies {
+    implementation project(':model')
     testImplementation platform('org.junit:junit-bom:5.10.0')
     testImplementation 'org.junit.jupiter:junit-jupiter'
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
@@ -32,7 +33,7 @@ sourceSets {
 }
 
 application {
-    mainClass = "ru.nsu.aeliseev2.task231.SnakeApp"
+    mainClass = "ru.nsu.aeliseev2.task231.app.SnakeApp"
 }
 
 run {
@@ -41,10 +42,4 @@ run {
 
 test {
     useJUnitPlatform()
-}
-
-jacocoTestReport {
-    reports {
-        xml.required = true
-    }
 }
