+++ Task_2_3_1/app/src/main/java/ru/nsu/aeliseev2/task231/app/SnakeApp.java
@@ -1,4 +1,4 @@
-package ru.nsu.aeliseev2.task231;
+package ru.nsu.aeliseev2.task231.app;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
