+++ Task_2_3_1/app/src/main/java/ru/nsu/aeliseev2/task231/app/drawing/FillRectTile.java
@@ -1,4 +1,4 @@
-package ru.nsu.aeliseev2.task231.drawing;
+package ru.nsu.aeliseev2.task231.app.drawing;
 
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Paint;
