+++ Task_2_3_1/app/src/main/java/ru/nsu/aeliseev2/task231/app/drawing/SnakeHeadTile.java
@@ -1,17 +1,17 @@
-package ru.nsu.aeliseev2.task231.drawing;
+package ru.nsu.aeliseev2.task231.app.drawing;
 
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Paint;
 
 /**
- * A tile that draws a part of the snake's body.
+ * A tile that draws the snake's head.
  *
  * @param primary   The primary paint.
  * @param secondary The secondary paint.
  * @param rotation  The rotation of this tile in degrees (counter-clockwise, starting from pointing
  *                  to the right).
  */
-public record SnakeBodyTile(Paint primary, Paint secondary, double rotation) implements Tile {
+public record SnakeHeadTile(Paint primary, Paint secondary, double rotation) implements Tile {
     /**
      * {@inheritDoc}
      */
@@ -21,10 +21,11 @@ public record SnakeBodyTile(Paint primary, Paint secondary, double rotation) imp
         context.translate(size / 2, size / 2);
         context.rotate(-rotation);
         context.translate(-size / 2, -size / 2);
-        context.setFill(secondary);
-        context.fillRect(size / 4, size / 4, size / 2, size / 2);
         context.setFill(primary);
-        context.fillRect(3 * size / 4, size / 4, size / 2, size / 2);
+        context.fillRect(size / 8, size / 8, 3 * size / 4, 3 * size / 4);
+        context.setFill(secondary);
+        context.fillOval(0.6 * size, 0.6 * size, size / 8, size / 8);
+        context.fillOval(0.6 * size, 0.3 * size, size / 8, size / 8);
         context.restore();
     }
 }
