+++ Task_2_3_1/model/src/main/java/ru/nsu/aeliseev2/task231/model/SnakeModel.java
@@ -1,0 +1,4 @@
+package ru.nsu.aeliseev2.task231.model;
+
+public class SnakeModel {
+}
