+++ Task_2_3_1/model/src/test/java/SnakeModelTests.java
@@ -1,0 +1,9 @@
+import org.junit.jupiter.api.Test;
+import ru.nsu.aeliseev2.task231.model.SnakeModel;
+
+class SnakeModelTests {
+    @Test
+    void creation() {
+        new SnakeModel();
+    }
+}
