+++ Task_2_3_1/settings.gradle
@@ -1,1 +1,3 @@
 rootProject.name = 'Task_2_3_1'
+include 'app'
+include 'model'
